refactor: extract independence into analyzer/independence.rs (#190)
Moves ModuleIndependence + compute_independence out of analyzer/mod.rs.
audit() now calls the helper instead of inlining the loop.
