+++ src/analyzer/mod.rs
@@ -11,6 +11,7 @@ use crate::core::{Dependency, DependencyDirection, Module};
 mod cohesion;
 mod critical_path;
 mod gravity_wells;
+mod independence;
 mod layers;
 mod red_flags;
 mod rules;
@@ -19,6 +20,7 @@ mod violation_age;
 pub use cohesion::{compute_cohesion, CohesionMetrics};
 pub use critical_path::compute_critical_path;
 pub use gravity_wells::{compute_gravity_wells, GravityWell};
+pub use independence::{compute_independence, ModuleIndependence};
 pub use layers::{check_layer_rules, LayerViolation};
 pub use red_flags::{compute_red_flags, RedFlag, RedFlagType};
 pub use rules::{check_dependency_rules, RuleViolation};
@@ -150,21 +152,6 @@ pub struct TopAction {
     pub impact: f64,
     /// Category: "circular", "layer", "rule", "cross-module", "hotspot".
     pub category: String,
-}
-
-/// Independence score for a top-level module/directory.
-#[derive(Debug, Clone)]
-pub struct ModuleIndependence {
-    /// Directory path.
-    pub dir: String,
-    /// Number of files in this module.
-    pub file_count: usize,
-    /// Dependencies where both source and target are within this module.
-    pub internal_deps: usize,
-    /// Dependencies where source is in this module but target is outside.
-    pub external_deps: usize,
-    /// Independence score: internal / (internal + external). Range 0.0 to 1.0.
-    pub independence: f64,
 }
 
 impl AuditResult {
@@ -1082,56 +1069,7 @@ pub fn audit(modules: &[Module], dependencies: &[Dependency]) -> AuditResult {
 
     let cohesion = compute_cohesion(modules, dependencies);
 
-    // Compute per-module independence (top-level directories)
-    let mut top_dirs: FxHashMap<String, FxHashSet<&str>> = FxHashMap::default();
-    for module in modules {
-        // Get the first path component as the top-level directory
-        let top = module
-            .path
-            .split('/')
-            .next()
-            .unwrap_or(&module.path)
-            .to_string();
-        // Only group if there's depth (file is not at root)
-        if module.path.contains('/') {
-            top_dirs.entry(top).or_default().insert(module.id.as_str());
-        }
-    }
-
-    let mut independence: Vec<ModuleIndependence> = top_dirs
-        .iter()
-        .filter(|(_, files)| files.len() >= 2)
-        .map(|(dir, files)| {
-            let internal = dependencies
-                .iter()
-                .filter(|d| {
-                    files.contains(d.from_module_id.as_str())
-                        && files.contains(d.to_module_id.as_str())
-                })
-                .count();
-            let external = dependencies
-                .iter()
-                .filter(|d| {
-                    files.contains(d.from_module_id.as_str())
-                        && !files.contains(d.to_module_id.as_str())
-                })
-                .count();
-            let total = internal + external;
-            let score = if total > 0 {
-                internal as f64 / total as f64
-            } else {
-                1.0
-            };
-            ModuleIndependence {
-                dir: dir.clone(),
-                file_count: files.len(),
-                internal_deps: internal,
-                external_deps: external,
-                independence: score,
-            }
-        })
-        .collect();
-    independence.sort_by(|a, b| a.independence.partial_cmp(&b.independence).unwrap());
+    let independence = compute_independence(modules, dependencies);
 
     // Calculate total XS: sum of weights for coupling + break_cost for circular
     let total_xs: usize = violations
